first

+++ bots/common_bot/models.py
@@ -93,7 +93,7 @@ class Unnecessary_word_filter(CreateTracker):
 class SuperSettings(CreateTracker):
     movie_channel_username= models.CharField(max_length=64, unique=True)
     movie_bot_username= models.CharField(max_length=64, unique=True)
-    trailer_channel_id= models.IntegerField(unique=True,default="-1002080046544")
+    trailer_channel_id= models.BigIntegerField(unique=True,default="-1002080046544")
 
     def __str__(self):
         return  "SuperSetttings"
